+++ src/MathGame.ConsoleApplication/Program.cs
@@ -1,5 +1,5 @@
 ﻿// -------------------------------------------------------------------------------------------------
-// MathGame.Program
+// MathGame.ConsoleApplication.Program
 // -------------------------------------------------------------------------------------------------
 // The insertion point of the console application.
 // -------------------------------------------------------------------------------------------------
