+++ src/MathGame.ConsoleApplication/Views/Menu.cs
@@ -1,5 +1,5 @@
 ﻿// -------------------------------------------------------------------------------------------------
-// MathGame.ConsoleApplication.Menu
+// MathGame.ConsoleApplication.Views.Menu
 // -------------------------------------------------------------------------------------------------
 // The menu for the console application.
 // -------------------------------------------------------------------------------------------------
