+++ src/MathGame/Models/Game.cs
@@ -1,7 +1,7 @@
 ﻿// -------------------------------------------------------------------------------------------------
 // MathGame.Models.Game
 // -------------------------------------------------------------------------------------------------
-// Data models that will be used in the console application.
+// Data models for the game data.
 // -------------------------------------------------------------------------------------------------
 using MathGame.Enums;
